+++ src/main/java/com/oliveira/budget/domain/product/RequestProductDTO.java
@@ -3,5 +3,5 @@ package com.oliveira.budget.domain.product;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-public record RequestProductDTO(UUID id, String name, String description, BigDecimal price) {
+public record RequestProductDTO(UUID id, String name, String description, Float price) {
 }
