+++ src/main/java/com/oliveira/budget/domain/product/UpdateProductDTO.java
@@ -2,5 +2,5 @@ package com.oliveira.budget.domain.product;
 
 import java.math.BigDecimal;
 
-public record UpdateProductDTO(String name, String description, BigDecimal price) {
+public record UpdateProductDTO(String name, String description, Float price) {
 }
